Separate url_decode from app.c
