+++ apps/sql-receptionist/include/utils/http.h
@@ -1,0 +1,1 @@
+extern char *url_decode(const char *src);
