+++ apps/sql-receptionist/src/utils/http.c
@@ -1,0 +1,30 @@
+/**
+ * Decodes URLs, like "test%20test" -> "test test"
+ * @param src The encoded URL to decode.
+ * @return A pointer to a series of characters representing the decoded URL.
+ */
+#include <stddef.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+char *url_decode(const char *src) {
+  size_t src_len = strlen(src);
+  char *decoded = malloc(src_len + 1);
+  size_t decoded_len = 0;
+
+  // decode %2x to hex
+  for (size_t i = 0; i < src_len; i++) {
+    if (src[i] == '%' && i + 2 < src_len) {
+      int hex_val;
+      sscanf(src + i + 1, "%2x", &hex_val);
+      decoded[decoded_len++] = hex_val;
+      i += 2;
+    } else {
+      decoded[decoded_len++] = src[i];
+    }
+  }
+
+  // add null terminator
+  decoded[decoded_len] = '\0';
+  return decoded;
+}
